Migrated to blueprint services, added initial Travis CI file.
