+++ pom.xml
@@ -37,15 +37,23 @@
 	</properties>
 	<dependencies>
 		<dependency>
+			<groupId>org.apache.aries.blueprint</groupId>
+			<artifactId>org.apache.aries.blueprint</artifactId>
+			<version>1.1.0</version>
+		</dependency>
+		<!--suppress NonOsgiMavenDependency-->
+		<dependency>
 			<groupId>org.apache.xmlrpc</groupId>
 			<artifactId>xmlrpc-server</artifactId>
 			<version>3.1.3</version>
 		</dependency>
+		<!--suppress NonOsgiMavenDependency-->
 		<dependency>
 			<groupId>org.robotframework</groupId>
 			<artifactId>javalib-core</artifactId>
 			<version>1.2.1</version>
 		</dependency>
+		<!--suppress NonOsgiMavenDependency-->
 		<dependency>
 			<groupId>org.apache.ws.commons.util</groupId>
 			<artifactId>ws-commons-util</artifactId>
@@ -88,6 +96,27 @@
 	<build>
 		<plugins>
 			<plugin>
+				<groupId>org.apache.maven.plugins</groupId>
+				<artifactId>maven-compiler-plugin</artifactId>
+				<version>2.3.2</version>
+				<configuration>
+					<source>1.7</source>
+					<target>1.7</target>
+				</configuration>
+			</plugin>
+			<plugin>
+				<groupId>org.apache.maven.plugins</groupId>
+				<artifactId>maven-javadoc-plugin</artifactId>
+				<version>2.8.1</version>
+				<executions>
+					<execution>
+						<goals>
+							<goal>jar</goal>
+						</goals>
+					</execution>
+				</executions>
+			</plugin>
+			<plugin>
 				<groupId>org.apache.felix</groupId>
 				<artifactId>maven-bundle-plugin</artifactId>
 				<extensions>true</extensions>
@@ -107,6 +136,9 @@
 							org.robotframework.*,
 							org.apache.xmlrpc.*
 						</Export-Package>
+						<Embed-Dependency>
+							*;artifactId=xmlrpc-server|javalib-core|ws-commons-util
+						</Embed-Dependency>
 					</instructions>
 				</configuration>
 			</plugin>
