+++ src/main/resources/OSGI-INF/blueprint/remote-server.xml
@@ -1,0 +1,11 @@
+<blueprint xmlns="http://www.osgi.org/xmlns/blueprint/v1.0.0">
+
+    <bean id="robotRemoteServer" class="org.robotframework.remoteserver.RemoteServerImpl" init-method="start"
+          destroy-method="stop">
+        <property name="port" value="8270"/>
+        <property name="host" value="0.0.0.0"/>
+    </bean>
+    <!-- Services -->
+    <service ref="robotRemoteServer" interface="org.robotframework.remoteserver.RemoteServer"/>
+</blueprint>
+
